# public/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'API G-10'
  description: "API de la SAE G-10. Pour l'utiliser, il faut se connecter avec un compte utilisateur et obtenir un token d'authentification. Ensuite, il faut ajouter ce token dans la section Authentification."
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api'
paths:
...
  /evenements:
    get:
      tags:
        - Evenements
      description: 'Récupère la liste des événements'
      operationId: 78fa335947f05269e6c380cab4abaa9c
      parameters:
        -
          name: type
          in: query
          description: "Type de l'événement"
          required: false
          schema:
            type: string
        -
          name: lieu
          in: query
          description: "Lieu de l'événement"
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Liste des événements'
        '404':
          description: 'Aucun événement trouvé'
    post:
      tags:
        - Evenements
      description: 'Crée un événement'
      operationId: dd762bcbe5f1c9a8964e8950856074a7
      parameters:
        -
          name: titre
          in: query
          description: "Titre de l'événement"
          required: true
          schema:
            type: string
        -
          name: description
          in: query
          description: "Description de l'événement"
          required: true
          schema:
            type: string
        -
          name: date_event
          in: query
          description: "Date de l'événement"
          required: true
          schema:
            type: string
        -
          name: lieu_id
          in: query
          description: "Lieu de l'événement"
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: 'Evénement créé avec succès'
        '422':
          description: 'Erreur de validation'
  '/evenements/{id}':
    get:
      tags:
        - Evenements
      description: 'Récupère un événement'
      operationId: 1bb284cafc3285decc946b651ac14928
      parameters:
        -
          name: id
          in: path
          description: "Identifiant unique de l'événement"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Evénement trouvé'
        '404':
          description: 'Evénement non trouvé'
    put:
      tags:
        - Evenements
      description: 'Mettre à jour un événement'
      operationId: ddc7d8c0e074964fddf00821812a7cc8
      parameters:
        -
          name: titre
          in: query
          description: "Titre de l'événement"
          required: false
          schema:
            type: string
        -
          name: description
          in: query
          description: "Description de l'événement"
          required: false
          schema:
            type: string
        -
          name: date_event
          in: query
          description: "Date de l'événement"
          required: false
          schema:
            type: string
        -
          name: lieu_id
          in: query
          description: "Lieu de l'événement"
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 'Evénement mis à jour'
        '404':
          description: 'Evénement non trouvé'
    post:
      tags:
        - Evenements
      description: 'Crée un prix pour un événement'
      operationId: 696aab9634515272d7fad7f7d9b72279
      parameters:
        -
          name: categorie
          in: query
          description: 'Catégorie du prix'
          required: true
          schema:
            type: string
        -
          name: nombre
          in: query
          description: 'Nombre de places'
          required: true
          schema:
            type: integer
        -
          name: valeur
          in: query
          description: 'Valeur du prix'
          required: true
          schema:
            type: number
      responses:
        '201':
          description: 'Prix créé avec succès'
        '422':
          description: 'Erreur de validation'
    delete:
      tags:
        - Evenements
      description: 'Supprime un événement'
      operationId: bb0b79e253d0510e1556f3ca79fad081
      parameters:
        -
          name: id
          in: path
          description: "Identifiant unique de l'événement"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Prix supprimé'
        '404':
          description: 'Prix non trouvé'
  '/evenements/{id}/updateParticipants':
    put:
      tags:
        - Evenements
      description: 'Récupère la liste des participants à un événement'
      operationId: 894d7f8ac0762469e2bb422f6d710d4b
      parameters:
        -
          name: id
          in: path
          description: "Identifiant unique de l'événement"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Liste des participants'
        '404':
          description: 'Aucun participant trouvé'
  '/evenements/{id}/prix':
    get:
      tags:
        - Evenements
      description: "Récupère la liste des prix d'un événement"
      operationId: 6454c72272b5d8d659513c8541a05a51
      parameters:
        -
          name: id
          in: path
          description: "Identifiant unique de l'événement"
          required: true
          schema:
            type: integer
            format: int64
        -
          name: categorie
          in: query
          description: 'Catégorie du prix'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Liste des prix'
        '404':
          description: 'Aucun prix trouvé'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer